Wyładowywanie hierarchii. poprawione błędy. Odświeżony python
git-svn-id: https://localhost:8443/svn/hm-mde-1@953 fca3399c-2208-499e-aaa1-2cf7fb07c25e

+++ trunk/src/edr/src/plugins/python/src/PythonLogic.cpp
@@ -40,7 +40,7 @@ void PythonLogic::run( const std::string& script)
     try {
         exec(code, mainNamespace);
         try {
-            int result = extract<int>(mainNamespace["result"]);
+            float result = extract<float>(mainNamespace["result"]);
             QMessageBox::information(nullptr, "Result", QString("Result : %1").arg(result));
         } catch (...) {
